handling MIN/MAX + 0 arg != error

+++ libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: gpaul <gpaul@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/26 15:33:12 by gpaul             #+#    #+#             */
-/*   Updated: 2021/06/30 16:12:16 by gpaul            ###   ########.fr       */
+/*   Updated: 2021/12/14 15:18:00 by gpaul            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long long int	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	re;
+	int				i;
+	int				sign;
+	long long int	re;
 
 	re = 0;
 	i = 0;
@@ -34,7 +34,7 @@ int	ft_atoi(const char *str)
 		re = re * 10 + (str[i++] - '0');
 	}
 	if (sign == 1)
-		return ((int) -re);
+		return (-re);
 	else
-		return ((int)re);
+		return (re);
 }
